+++ fdbclient/BackupAgentBase.actor.cpp
@@ -389,12 +389,14 @@ ACTOR static Future<Void> decodeBackupLogValue(Arena* arena,
 				Reference<AsyncVar<ClientDBInfo> const> dbInfo = cx->clientInfo;
 				try {
 					if (CLIENT_KNOBS->ENABLE_CONFIGURABLE_ENCRYPTION) {
-						TextAndHeaderCipherKeys cipherKeys = wait(getEncryptCipherKeys(
-						    dbInfo, logValue.configurableEncryptionHeader(), BlobCipherMetrics::RESTORE));
+						TextAndHeaderCipherKeys cipherKeys =
+						    wait(GetEncryptCipherKeys<ClientDBInfo>::getEncryptCipherKeys(
+						        dbInfo, logValue.configurableEncryptionHeader(), BlobCipherMetrics::RESTORE));
 						logValue = logValue.decrypt(cipherKeys, tempArena, BlobCipherMetrics::RESTORE);
 					} else {
-						TextAndHeaderCipherKeys cipherKeys = wait(
-						    getEncryptCipherKeys(dbInfo, *logValue.encryptionHeader(), BlobCipherMetrics::RESTORE));
+						TextAndHeaderCipherKeys cipherKeys =
+						    wait(GetEncryptCipherKeys<ClientDBInfo>::getEncryptCipherKeys(
+						        dbInfo, *logValue.encryptionHeader(), BlobCipherMetrics::RESTORE));
 						logValue = logValue.decrypt(cipherKeys, tempArena, BlobCipherMetrics::RESTORE);
 					}
 				} catch (Error& e) {
