+++ fdbclient/CommitProxyInterface.cpp
@@ -1,9 +1,11 @@
 #include "fdbclient/CommitProxyInterface.h"
 #include "fdbclient/CoordinationInterface.h"
+#include "fdbclient/GetEncryptCipherKeys_impl.actor.h"
 
 // Instantiate ClientDBInfo related tempates
 template class ReplyPromise<struct ClientDBInfo>;
 template class ReplyPromise<CachedSerialization<ClientDBInfo>>;
+template class GetEncryptCipherKeys<ClientDBInfo>;
 
 // Instantiate OpenDatabaseCoordRequest related templates
 template struct NetNotifiedQueue<OpenDatabaseCoordRequest, true>;
