+++ fdbclient/include/fdbclient/CommitProxyInterface.h
@@ -29,6 +29,7 @@
 #include "fdbclient/CommitTransaction.h"
 #include "fdbclient/EncryptKeyProxyInterface.h"
 #include "fdbclient/FDBTypes.h"
+#include "fdbclient/GetEncryptCipherKeys.actor.h"
 #include "fdbclient/GlobalConfig.h"
 #include "fdbclient/GrvProxyInterface.h"
 #include "fdbclient/IdempotencyId.actor.h"
@@ -748,5 +749,8 @@ struct GlobalConfigRefreshRequest {
 		serializer(ar, lastKnown, reply);
 	}
 };
+
+// Instantiated in CommitProxyInterface.cpp
+extern template class GetEncryptCipherKeys<ClientDBInfo>;
 
 #endif
