+++ fdbserver/RestoreLoader.actor.cpp
@@ -22,6 +22,7 @@
 // The RestoreLoader role starts with the restoreLoaderCore actor
 
 #include "fdbclient/BlobCipher.h"
+#include "fdbclient/CommitProxyInterface.h"
 #include "flow/UnitTest.h"
 #include "fdbclient/BackupContainer.h"
 #include "fdbclient/BackupAgent.actor.h"
@@ -376,8 +377,8 @@ ACTOR static Future<MutationRef> _decryptMutation(MutationRef mutation, Database
 	Reference<AsyncVar<ClientDBInfo> const> dbInfo = cx->clientInfo;
 	std::unordered_set<BlobCipherDetails> cipherDetails;
 	mutation.updateEncryptCipherDetails(cipherDetails);
-	std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>> getCipherKeysResult =
-	    wait(getEncryptCipherKeys(dbInfo, cipherDetails, BlobCipherMetrics::BACKUP));
+	std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>> getCipherKeysResult = wait(
+	    GetEncryptCipherKeys<ClientDBInfo>::getEncryptCipherKeys(dbInfo, cipherDetails, BlobCipherMetrics::BACKUP));
 	return mutation.decrypt(getCipherKeysResult, *arena, BlobCipherMetrics::BACKUP);
 }
