+++ fdbserver/include/fdbserver/ServerDBInfo.actor.h
@@ -126,6 +126,7 @@ struct GetServerDBInfoRequest {
 // Instantiated in worker.actor.cpp
 extern template class RequestStream<GetServerDBInfoRequest, false>;
 extern template struct NetNotifiedQueue<GetServerDBInfoRequest, false>;
+extern template class GetEncryptCipherKeys<ServerDBInfo>;
 
 ACTOR Future<Void> broadcastTxnRequest(TxnStateRequest req, int sendAmount, bool sendReply);
