+++ fdbserver/workloads/EncryptKeyProxyTest.actor.cpp
@@ -79,7 +79,8 @@ struct EncryptKeyProxyTestWorkload : TestWorkload {
 			domainIds.emplace(domainId);
 		}
 		std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>> latestCiphers =
-		    wait(getLatestEncryptCipherKeys(self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
+		    wait(GetEncryptCipherKeys<ServerDBInfo>::getLatestEncryptCipherKeys(
+		        self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
 
 		ASSERT_EQ(latestCiphers.size(), domainIds.size());
 
@@ -105,7 +106,8 @@ struct EncryptKeyProxyTestWorkload : TestWorkload {
 			domainIds.emplace(domainId);
 		}
 		std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>> latestCiphers =
-		    wait(getLatestEncryptCipherKeys(self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
+		    wait(GetEncryptCipherKeys<ServerDBInfo>::getLatestEncryptCipherKeys(
+		        self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
 
 		TraceEvent("SimPartialDomainIdCacheEnd");
 		return Void();
@@ -122,7 +124,8 @@ struct EncryptKeyProxyTestWorkload : TestWorkload {
 		}
 
 		std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>> latestCiphers =
-		    wait(getLatestEncryptCipherKeys(self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
+		    wait(GetEncryptCipherKeys<ServerDBInfo>::getLatestEncryptCipherKeys(
+		        self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
 		state std::vector<Reference<BlobCipherKey>> cipherKeysVec;
 		for (auto item : latestCiphers) {
 			cipherKeysVec.push_back(item.second);
@@ -149,7 +152,8 @@ struct EncryptKeyProxyTestWorkload : TestWorkload {
 			}
 
 			std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>> cipherKeys =
-			    wait(getEncryptCipherKeys(self->dbInfo, cipherDetails, BlobCipherMetrics::UsageType::TEST));
+			    wait(GetEncryptCipherKeys<ServerDBInfo>::getEncryptCipherKeys(
+			        self->dbInfo, cipherDetails, BlobCipherMetrics::UsageType::TEST));
 			// Ensure the sanity of the lookedup data
 			for (auto item : cipherKeys) {
 				bool found = false;
@@ -183,7 +187,8 @@ struct EncryptKeyProxyTestWorkload : TestWorkload {
 			}
 			domainIds.emplace(FDB_DEFAULT_ENCRYPT_DOMAIN_ID - 1);
 			std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>> res =
-			    wait(getLatestEncryptCipherKeys(self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
+			    wait(GetEncryptCipherKeys<ServerDBInfo>::getLatestEncryptCipherKeys(
+			        self->dbInfo, domainIds, BlobCipherMetrics::UsageType::TEST));
 			// BlobCipherKeyCache is 'empty'; fetching invalid cipher from KMS must through 'encrypt_key_not_found'
 			ASSERT(false);
 		} catch (Error& e) {
